requirements.txt
asn1crypto==1.4.0
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
beautifulsoup4==4.9.3
boto3==1.15.18
botocore==1.18.18
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==2.9.2
filelock==3.0.12
google-api-core==1.26.3
google-auth==1.30.0
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-cloud-translate==3.1.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.19
oauthlib==3.1.0
oscrypto==1.2.1
packaging==20.9
proto-plus==1.18.1
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.9.8
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-futures==1.0.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.4.3
soupsieve==2.2.1
urllib3==1.25.11
